Fix bug in drawing Ids

+++ norfair/drawing.py
@@ -20,7 +20,7 @@ def draw_predictions(frame, objects, radius=None, thickness=None, color=None, id
     if radius is None:
         radius = int(frame_scale)
     if id_size is None:
-        id_size = int(frame_scale / 5)
+        id_size = frame_scale / 5
     if thickness is None:
         thickness = int(frame_scale / 5)
     if color is None:
